requirements.txt
dj-database-url==0.5.0
gunicorn==20.1.0
psycopg2==2.9.1
Django==3.2.4
Pillow==8.2.0
asgiref==3.3.4
dj-config-url==0.1.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
pip==21.1.2
pytz==2021.1
setuptools==49.2.1
sqlparse==0.4.1
whitenoise==5.2.0